requirements.txt
semantic-kernel
asyncio
python-dotenv
azure-search-documents
fastapi
pandas
uvicorn
streamlit
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
